Use `install_items` to add networkd files instead
Adding all networkd files in /etc/systemd/network/ may not be desired
by all users. For starters there are network devices (wireguard) that
have private information that cannot be leaked in an initramfs.
Furthermore, some network configurations may not be feasible or even
desired at boot and need the full OS to set up, such as virtual
interfaces that depend on other services not in initramfs.

This commit adds a new file to example to show how to drop in
networkd config files into /etc/dracut.conf.d/ using install_items.
It also clarifies this change in the README and updates the ci to
follow the same configuration.